# docs/static/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "SOARCA API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/playbook/": {
            "get": {
                "description": "return all stored playbooks default limit:100",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playbook"
                ],
                "summary": "gets all the UUIDs for the stored playbooks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cacao.Playbook"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "submit a new playbook api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playbook"
                ],
                "summary": "submit playbook via the api",
                "parameters": [
                    {
                        "description": "playbook",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cacao.Playbook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cacao.Playbook"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/playbook/meta": {
            "get": {
                "description": "get playbook meta information for playbook",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playbook"
                ],
                "summary": "gets all the meta information for the stored playbooks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.PlaybookMeta"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/playbook/{id}": {
            "get": {
                "description": "get playbook by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playbook"
                ],
                "summary": "get CACAO playbook by its ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "playbook ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cacao.Playbook"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "update playbook by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playbook"
                ],
                "summary": "update playbook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "playbook Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "playbook",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cacao.Playbook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cacao.Playbook"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete playbook by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playbook"
                ],
                "summary": "delete playbook by Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "playbook ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
...
        "/trigger/playbook": {
            "post": {
                "description": "trigger playbook",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trigger"
                ],
                "summary": "trigger a playbook by supplying a cacao playbook payload",
                "parameters": [
                    {
                        "description": "execute playbook by payload",
                        "name": "playbook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cacao.Playbook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Execution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Error": {
            "type": "object",
            "required": [
                "message",
                "original-call",
                "status"
            ],
            "properties": {
                "downstream-call": {
                    "type": "string",
                    "example": "{\"some\" : \"json\"}"
                },
                "message": {
                    "type": "string",
                    "example": "missing argument in call"
                },
                "original-call": {
                    "type": "string",
                    "example": "/example/route"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "api.Execution": {
            "type": "object",
            "required": [
                "execution_id",
                "payload"
            ],
            "properties": {
                "execution_id": {
                    "type": "string",
                    "example": "2c855cd6-bbce-402f-a143-3d6eec346c08"
                },
                "payload": {
                    "type": "string",
                    "example": "playbook--0cec398c-db69-4f17-bde4-8ecbcc4a8879"
                }
            }
        },
...
        "api.PlaybookMeta": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "valid_from": {
                    "type": "string"
                },
                "valid_until": {
                    "type": "string"
                }
            }
        },
...
        "cacao.Addresses": {
            "type": "object",
            "additionalProperties": {
                "type": "array",
                "items": {
                    "type": "string"
                }
            }
        },
        "cacao.AgentTarget": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/cacao.Addresses"
                },
                "agent_target_extensions": {
                    "$ref": "#/definitions/cacao.Extensions"
                },
                "authentication_info": {
                    "type": "string"
                },
                "category": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "contact": {
                    "$ref": "#/definitions/cacao.Contact"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/cacao.CivicLocation"
                },
                "logical": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "sector": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "cacao.AgentTargets": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/cacao.AgentTarget"
            }
        },
        "cacao.AuthenticationInformation": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kms": {
                    "type": "boolean"
                },
                "kms_key_identifier": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "oauth_header": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "private_key": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "cacao.AuthenticationInformations": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/cacao.AuthenticationInformation"
            }
        },
...
        "cacao.Contact": {
            "type": "object",
            "properties": {
                "contact_details": {
                    "type": "string"
                },
                "email": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "phone": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "cacao.DataMarking": {
            "type": "object",
            "required": [
                "created",
                "created_by",
                "id",
                "type"
            ],
            "properties": {
                "affected_party_notifications": {
                    "type": "string"
                },
                "attribution": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "encrypt_in_transit": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "external_references": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cacao.ExternalReferences"
                    }
                },
                "id": {
                    "type": "string"
                },
                "iep_version": {
                    "type": "string"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "marking_extensions": {
                    "$ref": "#/definitions/cacao.Extensions"
                },
                "name": {
                    "type": "string"
                },
                "permitted_actions": {
                    "type": "string"
                },
                "revoked": {
                    "type": "boolean"
                },
                "start_date": {
                    "type": "string"
                },
                "statement": {
                    "type": "string"
                },
                "tlp": {
                    "type": "string"
                },
                "tlpv2_level": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "unmodified_resale": {
                    "type": "string"
                },
                "valid_from": {
                    "type": "string"
                },
                "valid_until": {
                    "type": "string"
                }
            }
        },
        "cacao.DataMarkings": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/cacao.DataMarking"
            }
        },
        "cacao.ExtensionDefinition": {
            "type": "object",
            "required": [
                "created_by",
                "schema",
                "type",
                "version"
            ],
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "external_references": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cacao.ExternalReferences"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "schema": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "cacao.ExtensionDefinitions": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/cacao.ExtensionDefinition"
            }
        },
        "cacao.Extensions": {
            "type": "object",
            "additionalProperties": true
        },
        "cacao.ExternalReferences": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "reference_id": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "cacao.Playbook": {
            "type": "object",
            "required": [
                "created",
                "created_by",
                "id",
                "modified",
                "name",
                "spec_version",
                "type",
                "workflow",
                "workflow_start"
            ],
            "properties": {
                "agent_definitions": {
                    "description": "Map of agent definitions used by the workflow steps",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cacao.AgentTargets"
                        }
                    ]
                },
                "authentication_info_definitions": {
                    "description": "Map of authentication information objects",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cacao.AuthenticationInformations"
                        }
                    ]
                },
                "created": {
                    "description": "Timestamp of the creation of the playbook",
                    "type": "string",
                    "example": "2024-01-01T09:00:00.000Z"
                },
                "created_by": {
                    "description": "UUID referring to identity",
                    "type": "string",
                    "example": "identity--96abab60-238a-44ff-8962-5806aa60cbce"
                },
                "data_marking_definitions": {
                    "description": "Map of datamarking definitions",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cacao.DataMarkings"
                        }
                    ]
                },
                "derived_from": {
                    "description": "Playbook id that this playbook is derived from",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"playbook--77c4c428-6304-4950-93ff-83c5224cb67a\"]"
                    ]
                },
                "description": {
                    "description": "A descriptive text to indicate what your playbook does",
                    "type": "string",
                    "example": "This is an example investigation playbook"
                },
                "extension_definitions": {
                    "description": "Map of extension definitions used by the workflow steps",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cacao.ExtensionDefinitions"
                        }
                    ]
                },
                "external_references": {
                    "description": "List of external reference objects",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cacao.ExternalReferences"
                    }
                },
                "id": {
                    "description": "Used by SOARCA so refer to the object while loading it from the database",
                    "type": "string",
                    "example": "playbook--77c4c428-6304-4950-93ff-83c5fd4cb67a"
                },
                "impact": {
                    "description": "A priority number ranging 0 - 100",
                    "type": "integer",
                    "example": 100
                },
                "labels": {
                    "description": "List of labels to label playbook",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "markings": {
                    "description": "List of datamarking identifiers",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[marking-statement--6424867b-0440-4885-bd0b-604d51786d06]"
                    ]
                },
                "modified": {
                    "description": "Timestamp of the last modification of the playbook",
                    "type": "string",
                    "example": "2024-01-01T09:00:00.000Z"
                },
                "name": {
                    "description": "An indicative name of the playbook",
                    "type": "string",
                    "example": "Investigation playbook"
                },
                "playbook_extensions": {
                    "description": "Map of extensions used by the playbook",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cacao.Extensions"
                        }
                    ]
                },
                "playbook_types": {
                    "description": "Should be of the CACAO playbook-type-ov",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "investigation"
                    ]
                },
                "playbook_variables": {
                    "description": "Map of variables that are global to the playbook",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cacao.Variables"
                        }
                    ]
                },
                "priority": {
                    "description": "A priority number ranging 0 - 100",
                    "type": "integer",
                    "example": 100
                },
                "severity": {
                    "description": "A priority number ranging 0 - 100",
                    "type": "integer",
                    "example": 100
                },
                "spec_version": {
                    "description": "Indicate the specification version cacao-2.0 is the only supported version at this time",
                    "type": "string",
                    "example": "cacao-2.0"
                },
                "target_definitions": {
                    "description": "Map of target definitions used by the workflow steps",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cacao.AgentTargets"
                        }
                    ]
                },
                "type": {
                    "description": "Must be playbook",
                    "type": "string",
                    "example": "playbook"
                },
                "valid_from": {
                    "description": "Timestamp from when the playbook is valid",
                    "type": "string",
                    "example": "2024-01-01T09:00:00.000Z"
                },
                "valid_until": {
                    "description": "Timestamp until when the playbook is valid",
                    "type": "string",
                    "example": "2124-01-01T09:00:00.000Z"
                },
                "workflow": {
                    "description": "Map of workflow steps keyed by the step id",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cacao.Workflow"
                        }
                    ]
                },
                "workflow_exception": {
                    "description": "Step that marks the actions that need to be taken when an exception occurs",
                    "type": "string",
                    "example": "end--37bea005-4a36-4a77-bd1f-79a6e4682a13"
                },
                "workflow_start": {
                    "description": "Start step of the playbook MUST be of step type START",
                    "type": "string",
                    "example": "start--07bea005-4a36-4a77-bd1f-79a6e4682a13"
                }
            }
        },
...
        "cacao.Variables": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/cacao.Variable"
            }
        },
        "cacao.Workflow": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/cacao.Step"
            }
        },
...
    }
}