# internal/transport/http/handlers/trigger/trigger_api.go
package trigger

import (
	"errors"
	"fmt"
	"io"
	"net/http"
	"reflect"
	"soarca/internal/logger"
	"soarca/internal/runs"
	apiError "soarca/internal/transport/http/handlers/error"
	"soarca/internal/transport/http/schema"
	"soarca/pkg/cacao"
	"soarca/internal/playbooks/decoder"

	"github.com/gin-gonic/gin"
)

type Empty struct{}

var log *logger.Log

type ITrigger interface {
	Execute(context *gin.Context)
}

func init() {
	log = logger.Logger(reflect.TypeOf(Empty{}).PkgPath(), logger.Info, "", logger.Json)
}

type TriggerHandler struct {
	runs runs.Runner
}

func NewTriggerHandler(runner runs.Runner) *TriggerHandler {
	return &TriggerHandler{runs: runner}
}

func (handler *TriggerHandler) ExecuteById(context *gin.Context) {
	log.Trace("received execute by ID")
	id := context.Param("id")
	var variables cacao.Variables
	if context.Request.Body != nil {
		jsonData, err := io.ReadAll(context.Request.Body)
		if err != nil {
			log.Trace("Playbook trigger has failed to decode request body")
			apiError.SendErrorResponse(context, http.StatusBadRequest, "Failed to decode request body", "POST /trigger/playbook/"+id, "")
			return
		}
		variables, err = runs.DecodeVariables(jsonData)
		if err != nil {
			log.Error(err)
			apiError.SendErrorResponse(context, http.StatusBadRequest, fmt.Sprintf("Cannot execute. reason: %s", err), "POST /trigger/playbook/"+id, "")
			return
		}
	}
	runID, err := handler.runs.StartByID(context.Request.Context(), id, variables)
	if err != nil {
		log.Error(err)
		var validationErr runs.ValidationError
		if errors.As(err, &validationErr) {
			apiError.SendErrorResponse(context, http.StatusBadRequest, fmt.Sprintf("Cannot execute. reason: %s", validationErr.Error()), "POST /trigger/playbook/"+id, "")
			return
		}
		apiError.SendErrorResponse(context, http.StatusRequestTimeout, err.Error(), "POST "+context.Request.URL.Path, "")
		return
	}
	context.JSON(http.StatusOK, api.RunStarted{
		RunId:      runID,
		PlaybookId: id,
	})
}

func (handler *TriggerHandler) Execute(context *gin.Context) {
	log.Trace("received execute with body")
	jsonData, err := io.ReadAll(context.Request.Body)
	if err != nil {
		log.Error("failed")
		apiError.SendErrorResponse(context, http.StatusBadRequest,
			"Failed to marshall json on server side",
			"POST /trigger/playbook", "")
		return
	}
	playbook := decoder.DecodeValidate(jsonData)
	if playbook == nil {
		log.Error("Failed to decode playbook")
		apiError.SendErrorResponse(context, http.StatusBadRequest,
			"Failed to decode playbook",
			"POST /trigger/playbook", "")
		return
	}

	runID, err := handler.runs.Start(context.Request.Context(), playbook, cacao.Variables{})
	if err != nil {
		log.Error(err)
		var validationErr runs.ValidationError
		if errors.As(err, &validationErr) {
			apiError.SendErrorResponse(context, http.StatusBadRequest, fmt.Sprintf("Cannot execute. reason: %s", validationErr.Error()), "POST /trigger/playbook", "")
			return
		}
		apiError.SendErrorResponse(context,
			http.StatusRequestTimeout,
			err.Error(),
			"POST "+context.Request.URL.Path, "")
		return
	}
	context.JSON(http.StatusOK,
		api.RunStarted{
			RunId:      runID,
			PlaybookId: playbook.ID,
		})
}